Adds CoreData persistence to ResultsSummary.

+++ Facial Exercises/Facial Exercises/CoreData/Exercise+Convenience.swift
@@ -16,7 +16,7 @@ import CoreData
 
 extension Exercise {
     
-    convenience init(type: String, identifier: UUID = UUID(), length: Double, score: Double?, context: NSManagedObjectContext = CoreDataStack.shared.mainContext) {
+    convenience init(type: String, identifier: UUID = UUID(), length: Float, score: Float?, context: NSManagedObjectContext = CoreDataStack.shared.mainContext) {
         self.init(context: context)
         self.identifier = identifier
         self.type = type
